Make it harder to import `native.dart` on the web.
Export an empty library if `dart:io` is not available.
(Using conditional export checking that `dart.library.io`
is either "false" or the empty string, since we don't have
negative conditions in conditional imports - which is deliberate and OK.)

That should make any import of `package:platform/native.dart`
not actually provide a `NativePlatform` class, instantly breaking
code which expects it to exist.